[Chapter 7] Feat: Fetch data for RevenueChart component
This commit imports the fetchRevenue function from data.ts and calls it inside the Page component to fetch data for the RevenueChart component. It also uncomment the RevenueChart component in both the Page component and the revenue-chart.tsx file. This change follows the data fetching section of Chapter 7 in the Next.js Learn course, which teaches how to fetch data for the dashboard overview page.