[Chapter 7] Feat: Fetch data for Card components
This commit imports the fetchCardData function from data.ts and calls it inside the Page component to fetch data for the Card components. The cards display the total amount of invoices collected, pending, and total, as well as the total number of customers. It also uncomment the Card components in the Page component. This change follows the practice section of Chapter 7 in the Next.js Learn course, which teaches how to fetch data for the dashboard overview page.