[Chapter 8] Feat: Make dashboard dynamic with unstable_noStore API
This commit adds the unstable_noStore API from next/cache to all the data fetching functions in data.ts. This prevents the responses from being cached and enables dynamic rendering for the dashboard. This change follows the static and dynamic rendering section of Chapter 8 in the Next.js Learn course, which teaches how to optimize your website for speed, scalability, and SEO.