[Chapter 9] Fix: Move loading.tsx and page.tsx to /(overview) folder
This commit moves the loading.tsx and page.tsx files from the dashboard folder to a new folder called /(overview) inside the dashboard folder. This fixes the loading skeleton bug that applied to the invoices and customers pages as well. This change follows the fixing the loading skeleton bug with route groups section of Chapter 9 in the Next.js Learn course, which teaches how to use route groups to control the loading state of different pages.