[Chapter 12] Feat: Create a Server Action and invoke it from the form
This commit adds a new Server Action in the lib directory and invokes it from the form. This change follows the "Create a Server Action" section of Chapter 12 in the Next.js Learn course, which teaches how to create a Server Action and invoke it from a form.

The changes include:
- Change 1: Created a new file named actions.ts in the lib directory.
- Change 2: Added the 'use server' directive at the top of the actions.ts file to mark all exported functions as server functions.
- Change 3: Imported the createInvoice function from the actions.ts file into the <Form> component.
- Change 4: Added an action attribute to the <form> element and called the createInvoice action.

This commit adds the ability for the form to create invoices by invoking the createInvoice Server Action. This improvement enhances the functionality of the form by allowing it to interact with the server and create invoices.